refactor(atlas-cli): extract pure CLI helpers to start-args.ts
The vitest for buildDaemonArgs / deriveHealthPort imported them from
start.tsx, which transitively pulls in `@atlas/client/v2`, credential
helpers, and other heavy modules. CI runs tests in parallel and the
load order can interact poorly with NATS-using tests in the suite
(`packages/jetstream/src/connect.test.ts` started failing after we
added vitest files that import start.tsx, even though the jetstream
file itself is byte-identical to main).

Move `buildDaemonArgs`, `deriveHealthPort`, and `StartArgs` to a new
`start-args.ts` — pure functions, no daemon-runtime imports. start.tsx
re-exports them so the rest of the CLI keeps its previously-stable
surface. The vitest now imports from start-args.ts directly.

Same pattern as the existing `health-listener-policy.ts` split that
sidestepped the broken daemon vitest harness.
